requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
bidict==0.21.4
charset-normalizer==2.0.12
dnspython==2.2.1
eventlet==0.33.1
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
idna-ssl==1.1.0
multidict==5.2.0
pkg-resources==0.0.0
python-engineio==4.3.2
python-socketio==5.6.0
six==1.16.0
typing-extensions==4.1.1
yarl==1.7.2
